+++ NEWS.md
@@ -3,6 +3,7 @@
 - Handle responses with no column information (fixes #49)
 - Add retries for GET and POST responses with error status codes
 - Skip test cases for ones that need locale modification if we cannot set the locale for the OS.
+- Adapt to changes in the upcoming dplyr and testthat versions.
 
 # RPresto 1.2.0